+++ lamp/Genie/Genie/FS/sys/mac/drive.cc
@@ -187,7 +187,7 @@ namespace Genie
 	template < class Accessor >
 	struct sys_mac_drive_N_Property : readonly_property
 	{
-		static const size_t fixed_size = Accessor::fixed_size;
+		static const int fixed_size = Accessor::fixed_size;
 		
 		static void get( plus::var_string& result, const vfs::node* that, bool binary )
 		{
